requirements.txt
Django>=4.0
psycopg2>=2.8
djangorestframework>=3.13.0
python-decouple==3.6
requests==2.28.1
